PR feedback - func rename and loc
